comment fixes: new testData folder and excluded from sonarcloud, log entire error, renamed utils.ts to index.ts
