+++ app/plugins/adp-backend/src/service/deliveryProgrammeRouter.ts
@@ -15,7 +15,7 @@ import {
   DeliveryProgramme,
   ProgrammeManager,
 } from '@internal/plugin-adp-common';
-import { checkForDuplicateTitle, getCurrentUsername } from '../utils/utils';
+import { checkForDuplicateTitle, getCurrentUsername } from '../utils/index';
 import { ProgrammeManagerStore } from '../deliveryProgramme/deliveryProgrammeManagerStore';
 import { Entity } from '@backstage/catalog-model';
 import {
@@ -56,9 +56,9 @@ export async function createProgrammeRouter(
       const data = await deliveryProgrammesStore.getAll();
       res.json(data);
     } catch (error) {
-      const errMsg = (error as Error).message;
-      logger.error('Error in retrieving delivery programmes: ', errMsg);
-      throw new InputError(errMsg);
+      const deliveryProgramError  = (error as Error);
+      logger.error('Error in retrieving delivery programmes: ', deliveryProgramError );
+      throw new InputError(deliveryProgramError.message);
     }
   });
 
@@ -67,9 +67,9 @@ export async function createProgrammeRouter(
       const data = await programmeManagersStore.getAll();
       res.json(data);
     } catch (error) {
-      const errMsg = (error as Error).message;
-      logger.error('Error in retrieving programme managers: ', errMsg);
-      throw new InputError(errMsg);
+      const deliveryProgramError = (error as Error);
+      logger.error('Error in retrieving programme managers: ', deliveryProgramError);
+      throw new InputError(deliveryProgramError.message);
     }
   });
 
@@ -84,9 +84,9 @@ export async function createProgrammeRouter(
         res.json(deliveryProgramme);
       }
     } catch (error) {
-      const errMsg = (error as Error).message;
-      logger.error('Error in retrieving delivery programme: ', errMsg);
-      throw new InputError(errMsg);
+      const deliveryProgramError = (error as Error);
+      logger.error('Error in retrieving delivery programme: ', deliveryProgramError);
+      throw new InputError(deliveryProgramError.message);
     }
   });
 
@@ -105,9 +105,9 @@ export async function createProgrammeRouter(
       });
       res.json(catalogApiResponse);
     } catch (error) {
-      const errMsg = (error as Error).message;
-      logger.error('Error in retrieving catalog entities: ', errMsg);
-      throw new InputError(errMsg);
+      const deliveryProgramError = (error as Error);
+      logger.error('Error in retrieving catalog entities: ', deliveryProgramError);
+      throw new InputError(deliveryProgramError.message);
     }
   });
 
@@ -162,9 +162,9 @@ export async function createProgrammeRouter(
         res.status(201).json(deliveryProgramme);
       }
     } catch (error) {
-      const errMsg = (error as Error).message;
-      logger.error('Error in creating a delivery programme: ', errMsg);
-      throw new InputError(errMsg);
+      const deliveryProgramError = (error as Error);
+      logger.error('Error in creating a delivery programme: ', deliveryProgramError);
+      throw new InputError(deliveryProgramError.message);
     }
   });
 
@@ -264,9 +264,9 @@ export async function createProgrammeRouter(
       }
       res.status(200).json(deliveryProgramme);
     } catch (error) {
-      const errMsg = (error as Error).message;
-      logger.error('Error in updating a delivery project: ', errMsg);
-      throw new InputError(errMsg);
+      const deliveryProgramError = (error as Error);
+      logger.error('Error in updating a delivery project: ', deliveryProgramError);
+      throw new InputError(deliveryProgramError.message);
     }
   });
