+++ app/plugins/adp-backend/src/testData/projectTestData.ts
@@ -1,0 +1,20 @@
+export const expectedProjectData = {
+  title: 'Test title',
+  alias: 'Test alias',
+  description: 'Test description',
+  finance_code: 'Test finance_code',
+  delivery_project_code: 'Test delivery_project_code',
+  url: 'Test url',
+  ado_project: 'Test ado_project'
+};
+
+export const expectedProjectDataWithName = {
+  title: 'Test title',
+  alias: 'Test alias',
+  description: 'Test description',
+  finance_code: 'Test finance_code',
+  delivery_project_code: 'Test delivery_project_code',
+  url: 'Test url',
+  ado_project: 'Test ado_project',
+  name: 'Test name'
+};
